Extract hook for class-card data
