+++ frontend/src/components/teacher/student-management/classroom-summary/useClassDataQuery.ts
@@ -1,0 +1,48 @@
+import { useContext, useMemo } from "react";
+import { useQuery } from "@apollo/client";
+
+import { GET_CLASSES_BY_TEACHER } from "../../../../APIClients/queries/ClassQueries";
+import type { ClassResponse } from "../../../../APIClients/types/ClassClientTypes";
+import AuthContext from "../../../../contexts/AuthContext";
+import { getSessionStatus } from "../../../../utils/TestSessionUtils";
+
+const useClassDataQuery = (limit?: number) => {
+  const { authenticatedUser } = useContext(AuthContext);
+  const { id: teacherId } = authenticatedUser ?? {};
+
+  const { loading, error, data } = useQuery<{
+    classesByTeacher: ClassResponse[];
+  }>(GET_CLASSES_BY_TEACHER, {
+    fetchPolicy: "cache-and-network",
+    variables: { teacherId, limit },
+    skip: !teacherId,
+  });
+
+  const classCards = useMemo(() => {
+    const now = new Date();
+    return data?.classesByTeacher.map(
+      ({ testSessions, students, ...classCard }) => {
+        let activeAssessments = 0;
+        testSessions.forEach((session) => {
+          if (
+            getSessionStatus(session.startDate, session.endDate, now) ===
+            "active"
+          ) {
+            activeAssessments += 1;
+          }
+        });
+
+        return {
+          ...classCard,
+          activeAssessments,
+          assessmentCount: testSessions.length,
+          studentCount: students.length,
+        };
+      },
+    );
+  }, [data]);
+
+  return { loading, error, data: classCards };
+};
+
+export default useClassDataQuery;
